+++ sys/dosmem.h
@@ -10,12 +10,12 @@
 # include "mint/mem.h"
 
 long _cdecl sys_m_addalt (long start, long size);
-long _cdecl sys_m_xalloc (long size, int mode);
+long _cdecl sys_m_xalloc (long size, short mode);
 long _cdecl sys_m_alloc (long size);
 long _cdecl sys_m_free (long block);
-long _cdecl sys_m_shrink (int dummy, unsigned long block, long size);
-long _cdecl sys_m_validate (int pid, unsigned long addr, long size, long *flags);
-long _cdecl sys_m_access (unsigned long addr, long size, int mode);
+long _cdecl sys_m_shrink (short dummy, unsigned long block, long size);
+long _cdecl sys_m_validate (short pid, unsigned long addr, long size, long *flags);
+long _cdecl sys_m_access (unsigned long addr, long size, short mode);
 
 
 # endif /* _dosmem_h */
